// src/api/accounts/models/user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * Document Mapping class.  Fields:  - `id` - MongoDB document ObjectID \"_id\" field. Mapped to the PydanticObjectId class  Inherited from:  - Pydantic BaseModel - [UpdateMethods](https://roman-right.github.io/beanie/api/interfaces/#aggregatemethods)
 * @export
 * @interface User
 */
export interface User {
  /**
   *
   * @type {string}
   * @memberof User
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof User
   */
  user_id: string;
  /**
   * The first name of the user
   * @type {string}
   * @memberof User
   */
  firstName: string;
  /**
   * The users last name
   * @type {string}
   * @memberof User
   */
  lastName: string;
  /**
   * The email of the user
   * @type {string}
   * @memberof User
   */
  email: string;
  /**
   * The mobile number must be a valid phone number
   * @type {string}
   * @memberof User
   */
  mobileNumber?: string;
  /**
   * The description of the user
   * @type {string}
   * @memberof User
   */
  description?: string;
  /**
   * The list of the user's favorite items
   * @type {Array<string>}
   * @memberof User
   */
  favorites?: Array<string>;
  /**
   * The user's avatar
   * @type {string}
   * @memberof User
   */
  avatar?: string;
  /**
   * The date the user was created
   * @type {Date}
   * @memberof User
   */
  createdAt?: Date;
  /**
   * The date the user was updated
   * @type {Date}
   * @memberof User
   */
  updatedAt?: Date;
}